+++ metadata.rb
@@ -9,7 +9,8 @@ chef_version '>= 14'
 version '7.1.0'
 
 depends 'ark', '>= 2.0.2'
+depends 'chocolatey', '>= 3.0'
 
-%w(debian ubuntu centos redhat scientific oracle amazon smartos mac_os_x opensuseleap suse).each do |os|
+%w(debian ubuntu centos redhat scientific oracle amazon smartos mac_os_x opensuseleap suse windows).each do |os|
   supports os
 end