+++ test/cookbooks/test/recipes/chocolatey.rb
@@ -1,0 +1,3 @@
+node.default['nodejs']['install_method'] = 'chocolatey'
+include_recipe 'nodejs::default'
+include_recipe 'test::resource_win'