+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Verify the URI of installed packages to help determine if a good URI has been installed
 - Add tests that verify npm-package installed packages
 - Get the example test cookbook working
+- Add support for installing node on windows
 
 ## 7.1.0 (2020-10-01)